tests-entry.js replaces unit-tests.js, integration-tests.js, storybook-tests.js
