--- apps/test/storybook-tests.js
@@ -1,6 +1,0 @@
-import '@babel/polyfill/noConflict';
-import 'whatwg-fetch';
-
-describe('storybook tests', function () {
-  require('./storybook/renderStoriesTest');
-});
